+++ src/analysis/processing/qgsalgorithmintersection.cpp
@@ -113,8 +113,8 @@ QVariantMap QgsIntersectionAlgorithm::processAlgorithm( const QVariantMap &param
   QVariantMap outputs;
   outputs.insert( QStringLiteral( "OUTPUT" ), dest );
 
-  long count = 0;
-  const long total = sourceA->featureCount();
+  long long count = 0;
+  const long long total = sourceA->featureCount();
 
   QgsGeometryParameters geometryParameters;
   if ( parameters.value( QStringLiteral( "GRID_SIZE" ) ).isValid() )
